test(join_columns): increase rigor of test data set with additional column
Previous tests were passed even though the join_columns() function was not behaving as expected: this was because I only had two input columns in the test, `id` and `AGE_bucket`. I've now added an additional neutral data column that has to be carried forward, unchanged, in the output for the test to be passed.