feat(fig02, fig02a): define helper function for table making
Use this function to more neatly produce figures 2 and 2a, and use existing data in db/ipums-processed.duckdb to avoid re-bucketing ages. Only pitfall of this approach is the brittleness of the age bucket definition and the need for these buckets to have been assigned upstream in the workflow. This is an area for future improvment.